+++ SysID Addon/seeq_sysid/__init__.py
@@ -1,1 +1,4 @@
 from . import gui, model
+from ._version import __version__
+
+__all__ = ['__version__']